# src/main/webapp/WEB-INF/swagger.yaml
...
swagger: "2.0"
info:
  description: "This API provides access to the internal PROXY for internal REST services"
  version: "3.0.0"
  title: "The ParaBank services proxy REST API"
  termsOfService: "/parabank/about.htm"
  contact:
    name: "ParaBank Web Administrator"
    url: "http://www.parasoft.com"
    email: "webadmin@parabank.parasoft.com"
  license:
    name: "Apache 2.0 License"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "replace-host-name"
basePath: "/parabank/services_proxy/bank"
...
schemes:
- "http"
produces:
- "application/json"
paths:
...
  /customers/{customerId}/loanRequests:
    get:
      tags:
      - "Loans"
      summary: "Get customer loan request history"
      operationId: "getLoanRequests"
      produces:
      - "application/json"
      parameters:
      - name: "customerId"
        in: "path"
        description: "Customer's id"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Loan request history, or an empty list when no requests exist"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/LoanRequestHistory"
        404:
          description: "Customer not found"
...
  /customers/{customerId}:
    get:
      tags:
      - "Customers"
      summary: "Get Customer Details"
      description: ""
      operationId: "getCustomer_1"
      produces:
      - "application/json"
      parameters:
      - name: "customerId"
        in: "path"
        description: "Customer's id"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Customer"
...
definitions:
...
  Address:
    type: "object"
    properties:
      street:
        type: "string"
      city:
        type: "string"
      state:
        type: "string"
      zipCode:
        type: "string"
...
  Customer:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      firstName:
        type: "string"
      lastName:
        type: "string"
      address:
        $ref: "#/definitions/Address"
      phoneNumber:
        type: "string"
      ssn:
        type: "string"
    xml:
      name: "customer"
...
  LoanRequestHistory:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      customerId:
        type: "integer"
        format: "int32"
      requestDate:
        type: "string"
        format: "date-time"
      availableFunds:
        type: "number"
      loanAmount:
        type: "number"
      downPayment:
        type: "number"
      approved:
        type: "boolean"
      responseDate:
        type: "string"
        format: "date-time"
      loanAccountId:
        type: "integer"
        format: "int32"
      providerName:
        type: "string"
      message:
        type: "string"
...